+++ src/Admin/Authorization/ChiefPasswordBroker.php
@@ -6,12 +6,8 @@ namespace Thinktomorrow\Chief\Admin\Authorization;
 
 use Illuminate\Auth\Passwords\PasswordBroker;
 
-/** @inheritDoc */
+/** {@inheritDoc} */
 class ChiefPasswordBroker extends PasswordBroker
 {
-    const RESET_LINK_SENT = 'chief::passwords.sent';
-    const PASSWORD_RESET = 'chief::passwords.reset';
-    const INVALID_USER = 'chief::passwords.user';
-    const INVALID_PASSWORD = 'chief::passwords.password';
-    const INVALID_TOKEN = 'chief::passwords.token';
+
 }